Extract Actions to a separate file
